Refactoring executors

+++ app/src/main/java/com/yuriy/openradio/shared/model/storage/drive/GoogleDriveHelper.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2017-2020 The "Open Radio" Project. Author: Chernyshov Yuriy
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.yuriy.openradio.shared.model.storage.drive;
 
 import androidx.annotation.NonNull;
@@ -9,14 +25,14 @@ import com.google.api.client.http.ByteArrayContent;
 import com.google.api.services.drive.Drive;
 import com.google.api.services.drive.model.File;
 import com.google.api.services.drive.model.FileList;
+import com.yuriy.openradio.shared.utils.AppLogger;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.Collections;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
+import java.util.concurrent.ExecutorService;
 
 /**
  * A utility for performing read/write operations on Drive files via the REST API.
@@ -25,7 +41,6 @@ public final class GoogleDriveHelper {
 
     private static final String MIME_TYPE_TEXT = "text/plain";
     private static final String MIME_TYPE_FOLDER = "application/vnd.google-apps.folder";
-    private final Executor mExecutor = Executors.newSingleThreadExecutor();
     private final Drive mDriveService;
 
     public GoogleDriveHelper(@NonNull final Drive driveService) {
@@ -36,9 +51,14 @@ public final class GoogleDriveHelper {
     /**
      * Creates a text file in the user's My Drive folder and returns its file ID.
      */
-    public final Task<String> createFile(final String folderId, @NonNull final String name,
+    public final Task<String> createFile(@NonNull final ExecutorService executorService,
+                                         final String folderId, @NonNull final String name,
                                          @NonNull final String content) {
-        return Tasks.call(mExecutor, () -> {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle create file requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> {
             final File metadata = new File()
                     .setParents(Collections.singletonList(folderId))
                     .setMimeType(MIME_TYPE_TEXT)
@@ -58,8 +78,13 @@ public final class GoogleDriveHelper {
      * @param name
      * @return
      */
-    public final Task<String> createFolder(@NonNull final String name) {
-        return Tasks.call(mExecutor, () -> {
+    public final Task<String> createFolder(@NonNull final ExecutorService executorService,
+                                           @NonNull final String name) {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle create folder requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> {
                     final File metadata = new File()
                             .setMimeType(MIME_TYPE_FOLDER)
                             .setName(name);
@@ -69,15 +94,19 @@ public final class GoogleDriveHelper {
                     return folder.getId();
                 }
         );
-
     }
 
     /**
      * Opens the file identified by {@code fileId} and returns a {@link Pair} of its name and
      * contents.
      */
-    public final Task<Pair<String, String>> readFile(@NonNull final String fileId) {
-        return Tasks.call(mExecutor, () -> {
+    public final Task<Pair<String, String>> readFile(@NonNull final ExecutorService executorService,
+                                                     @NonNull final String fileId) {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle read file requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> {
                     // Retrieve the metadata as a File object.
                     final File metadata = mDriveService.files().get(fileId).execute();
                     final String name = metadata.getName();
@@ -99,10 +128,15 @@ public final class GoogleDriveHelper {
      * Updates the file identified by {@code fileId} with the given {@code name} and {@code
      * content}.
      */
-    public final Task<Void> saveFile(@NonNull final String fileId,
+    public final Task<Void> saveFile(@NonNull final ExecutorService executorService,
+                                     @NonNull final String fileId,
                                      @NonNull final String name,
                                      @NonNull final String content) {
-        return Tasks.call(mExecutor, () -> {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle save file requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> {
             // Create a File containing any metadata changes.
             File metadata = new File().setName(name);
             // Convert content to an AbstractInputStreamContent instance.
@@ -113,10 +147,13 @@ public final class GoogleDriveHelper {
         });
     }
 
-    public final Task<Void> deleteFile(@NonNull final String fileId) {
-        return Tasks.call(
-                mExecutor,
-                () -> {
+    public final Task<Void> deleteFile(@NonNull final ExecutorService executorService,
+                                       @NonNull final String fileId) {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle delete file requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> {
                     // Delete file with specified id..
                     mDriveService.files().delete(fileId).execute();
                     return null;
@@ -130,10 +167,12 @@ public final class GoogleDriveHelper {
      * <p>The returned list will only contain folder visible to this app, i.e. those which were
      * created by this app.</p>
      */
-    public final Task<FileList> queryFolder(final String name) {
-        return Tasks.call(
-                mExecutor,
-                () -> mDriveService.files().list()
+    public final Task<FileList> queryFolder(@NonNull final ExecutorService executorService, final String name) {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle query folder requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> mDriveService.files().list()
                         .setSpaces("drive")
                         .setQ("mimeType='" + MIME_TYPE_FOLDER + "' and trashed=false and name='" + name + "'")
                         .execute()
@@ -146,10 +185,12 @@ public final class GoogleDriveHelper {
      * <p>The returned list will only contain folder visible to this app, i.e. those which were
      * created by this app.</p>
      */
-    public final Task<FileList> queryFile(final String fileName) {
-        return Tasks.call(
-                mExecutor,
-                () -> mDriveService.files().list()
+    public final Task<FileList> queryFile(@NonNull final ExecutorService executorService, final String fileName) {
+        if (executorService.isShutdown()) {
+            AppLogger.e("Executor is terminated, can't handle query file requests");
+            return Tasks.forCanceled();
+        }
+        return Tasks.call(executorService, () -> mDriveService.files().list()
                         .setSpaces("drive")
                         .setQ("mimeType='" + MIME_TYPE_TEXT + "' and trashed=false and name='" + fileName + "'")
                         .execute()
